VNIC: Fix for HW bug 21858

Changes:
1. HW bug 21858 mandates Qset to be enabled before any of the queues are enabled.
2. Also changed queue configuration sequence
3. Forgot to add that in the previous commit CQ_THRESH is disabled
   and now only CQ-TIMER_THRESH is being used.
   This is due to unexpected way of threshold implementation in hardware where
   it won't generate interrupt if the Qcount is already passed thresh value.
   Even with the timer thresh currently it seems to work just like a timer interrupt
   rather than a interrupt after expiry with some valid qcount.
   Need to varify this again.

+++ drivers/net/ethernet/cavium/thunder/nicvf_queues.h
@@ -260,7 +260,7 @@ struct rcv_queue {
 
 struct cmp_queue {
 	bool		enable;
-	uint8_t		intr_timer_thresh;
+	uint16_t	intr_timer_thresh;
 	uint16_t	thresh;
 	spinlock_t	cq_lock;  /* lock to serialize processing CQEs */
 	void		*desc;
